requirement.txt
fastapi
uvicorn
python-multipart
pillow
numpy
pytest
mlflow
dvc
requests
google-cloud-storage
